move evaluations and pipelines
