+++ examples/MotorImagery/two_class_motor_imagery.py
@@ -1,4 +1,4 @@
-from moabb.contexts.evaluations import WithinSessionEvaluation
+from moabb.evaluations import WithinSessionEvaluation
 from moabb.contexts.motor_imagery import LeftRightImagery
 from pyriemann.estimation import Covariances
 from pyriemann.spatialfilters import CSP
