+++ moabb/evaluations/__init__.py
@@ -1,0 +1,1 @@
+from moabb.evaluations.evaluations import *
