+++ moabb/tests/contexts.py
@@ -1,4 +1,4 @@
-from moabb.contexts import evaluations as ev
+from moabb.evaluations import evaluations as ev
 from moabb.datasets.bnci import BNCI2014001
 from moabb.viz import Results
 from moabb.contexts.motor_imagery import LeftRightImagery
@@ -16,6 +16,7 @@ pipelines = OrderedDict()
 pipelines['C'] = make_pipeline(Covariances('oas'), CSP(8), LDA())
 d = BNCI2014001()
 d.selected_events = {k: d.event_id[k] for k in ['left_hand', 'right_hand']}
+
 
 class Test_CrossSess(unittest.TestCase):
     '''This is actually integration testing but I don't know how to do this
